Fix import issues with User model
