+++ packages/components/src/internal/app/models.ts
@@ -5,7 +5,9 @@
 import { Record } from 'immutable';
 import { ActionURL, getServerContext } from '@labkey/api';
 
-import { Container, User } from '../components/base/models/model';
+import { User } from '../..'
+
+import { Container } from '../components/base/models/model';
 
 const user = new User(getServerContext().user);
