+++ packages/components/src/internal/components/forms/PageDetailHeader.spec.tsx
@@ -17,7 +17,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { mount } from 'enzyme';
 
-import { User } from '../base/models/model';
+import { User } from '../../..';
 
 import { PageDetailHeader } from './PageDetailHeader';
