+++ packages/components/src/internal/components/forms/PageDetailHeader.tsx
@@ -17,7 +17,7 @@ import React, { PureComponent, ReactNode } from 'react';
 import { PermissionTypes } from '@labkey/api';
 
 import { hasAllPermissions } from '../../util/utils';
-import { User } from '../base/models/model';
+import { User } from '../../..';
 import { SVGIcon } from '../base/SVGIcon';
 
 import { FieldEditorOverlay, FieldEditorOverlayProps } from './FieldEditorOverlay';
