+++ packages/components/src/test/data/users.ts
@@ -1,6 +1,6 @@
 import { List } from 'immutable';
 
-import { User } from '../../internal/components/base/models/model';
+import { User } from '../..';
 
 export const TEST_USER_GUEST = new User({
     canDelete: false,
